Fixes: Review correction implementation
