+++ app/src/main/java/com/udacity/gradle/builditbigger/EndpointsAsyncTask.java
@@ -46,7 +46,8 @@ class EndpointsAsyncTask extends AsyncTask<Void, Void, String> {
         try {
             return myApiService.tellAJoke().execute().getData();
         } catch (IOException e) {
-            return e.getMessage();
+            e.printStackTrace();
+            return "";
         }
     }
